pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "colorice"
version = "0.1.0"
description = "Generate terminal color schemes from wallpaper images"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0",
    "scikit-learn>=1.3",
]

[project.scripts]
colorice = "colorice.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
